Mass Class Renaming
I Basically, Renamed every class in the game. turns out you arent supposed to name classes 'class_name' but should use ClassName as the underscore and break stuff.

so i went through and fixed all of that. Also added a button reset before connecting signals and a slight adjustment to the dialog UI. now on to the importer!